requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
attrs==24.2.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
paho-mqtt==2.1.0
propcache==0.2.1
yarl==1.18.3
